requirements.txt
python-ldap
pyyaml
voluptuous
yamllint
